refactor: move get_current_repository into git module
