+++ lua/gh-actions/providers/github/rest/_api.lua
@@ -1,33 +1,5 @@
 local M = {}
 
----@param str string
----@return string
-local function strip_git_suffix(str)
-  if str:sub(-4) == '.git' then
-    return str:sub(1, -5)
-  end
-
-  return str
-end
-
-function M.get_current_repository()
-  local job = require('plenary.job')
-  local origin_url_job = job:new {
-    command = 'git',
-    args = {
-      'config',
-      '--get',
-      'remote.origin.url',
-    },
-  }
-
-  origin_url_job:sync()
-
-  local origin_url = table.concat(origin_url_job:result(), '')
-
-  return strip_git_suffix(origin_url):match('([^@/:]+)[:/]([^/]+/[^/]+)$')
-end
-
 ---@param cmd? string
 ---@param server? string
 ---@return string|nil
